Second Commit
When the first version of the "PlayerController" was tested I noticed somethings. The players "jumpForce" was too high meaning i'd have to decrease it, or increase gravity on "Rigidbody 2D". Another thing I noticed was when going towards the left the player would stay facing to the right meaning I will have to add a flip feature to my "PlayerController" script. Once I fixed these problems, I did this by increasing the gravity on the player from 1 to 3. I noticed the player would fall over if moving too fast so I froze the "z" axis on the "Rigidbody 2D".