Fourth Commit
I noticed that when designing the "EnemyPatrol" script that once the player entered the dead zone of the camera the player would disappear off screen so I decided to add a "Cinemachine 2D Camera" to follow the player. Smoothing was also added to smooth out the camera as it moved.
Then added the "Cinemachine 2D Camera" as a child file to the main camera to tidy up. Also background colour was changed to lighter blue to repersent a sky blue.