Fifth Commit
Player will now respawn when player falls off the platform into the "Void", which is just a square sprite which the sprite render has been disabled so that the void isnt visable. The square sprite can be stretched along the level allowing the player to respawn when the player falls off the platform or if a glitch may occurs. The square sprite has a "Box Collider 2D" added "is triggered" turned on so that the "FallenOff" script will work. A "Respawn Point" was added so that the players position will now be instantly moved to the location of the "Respawn Point". A delay may be added to torture the player for dying as well as a respawn screen or game over screen.