Sixth Commit
I will now be creating "Boarders" which will be placed at the start of the level so that the player cannot go off the map into the void and to also keep the player on track in the right direction.  A boarder on the far right may be placed so that if i want the player to head upwards or downwards instead of left or right. A "Boarder" on the far right will be added to avoid errors for when I create a "NextLevel"l script. The "Boarders" are simple rectangle sprites with sprite renders disabled and "Box Collider 2D" added to the sprite. This is so the "Boarder" will not be visable and will still collide with the player stoping them. Tags were also added for testing purposes.