Seventh Commit
After my "Boarders" were created I realised that I would need to add a way of allowing the player to go to the next level. I realised I could do this with increasing the "buildIndex" and adding this code to a "Box Collider 2D" with "isTriggered" enabled and with the sprite render disabled so that its not visable. A "Test Scene" was also created to test the increase of the "buildIndex" with the "NextLevel" script. A "PerviousLevel" may be added to give the game a more openworld map.