Eighth Commit
This time I have decided to add heath to the player so that the player can take damage from enemies and any other objects that can deal damage. I decided to do hearts at the top of the screen so 3 "Full Hearts" to 0 "Empty Hearts", rather than a bar that will be decreased from 100% to 0% as this will give a more simple feeling across the game. I first added in the two "Heart" sprites and made them 32x32 and splitted them to devide them between Full Hearts" and "Empty Hearts".  Then added a "Canvas" and set it to scale with the screen aswell as setting the resolution to 1920x1080 or "HD". Then I added a Image as a child file under the "Canvas" and then set the "Full Heart" as the image and then duplicated it three times as the player will have three hearts when spawning in, not to little not to much either. This will change through out the game depending if the player takes damage and will be reset when player falls off into the "Void" or loses all of their hearts and then respawns with full hearts again. I then positioned them where i wanted on the screen then added a "PlayerHealth" script. I made so that the number of hearts stayed at 3 but health could decrease less than 3 but could not exceed it. I then refered to the "UnityEngine.UI to change the sprites during the game, Unity forms helped me alot through this. I also made health a public variable for testing purposes. After doing this I added it to the player and placed in my "Full Heart" sprite and "Empty Heart" sprite into the script. This worked very well and all I need to do now is taking damage and dying and respawning. I will do this in the next commit.