Ninth Commit
Now I will be coding for the "EnemyAttack" script I will use a private void so that when the enemy collides with the player it will then refer to a void the "PlayerHealth" script to take away a heart. Then I realised I would have to make a if player has equal or less health then zero then player will die. A void was made for if the player dies and then I also added a respawn void for when the player die then the player will respawn at the start of the level. I used the same script as the "FallenOff" script which will move the players position to the empty named respawns location. An extra platfrom was added to test the "PlayerHealth" script. The "Boarder Left" was moved further left to allow the player to move more. While testing  I noticed that the player would be pushed back by dino then i realised I need to set on the dino "Box Collider 2D" to istriggered. All scripts worked well but I thought I will need to add a way to make the player not take damage straight after taking damage otherwise the player will die almost instantly. I will do this by making a bool called "invincible" and when this bool equals true the players "Box Collider 2D" will disable. This means player will no longer be able to take damage. Adding the player "Box Collider 2D" didnt help so I used the dino "Box Collider 2D" and this worked well. I used Invoke for the first time and this was used to delay the amount of time the player is invincible for and when it will be turned off so that the player is not invincible for ever, I set this for only a few seconds.