Tenth Commit
A menu is needed to start the game off. I did this by making a new scene and adding a canvas set to scale to screen and default resolution at 1920x1080. I also added two buttons, a "Play" button and a "Quit" button and added it as a child file under an empty called "Buttons" a UI image was also added to set a image as a background,   but I may add the player and make it move in the background as a live background. A game title was also added set as "Game Title" I will decide on the game name later on. "Text Mesh Pro" extension was added to add more fonts to the buttons and title. Then i added a "Menu" script and added it to the empty called "Buttons". In the script I make sure to add "UnityEngine.SceneManagement" so that I can increase the "buildIndex" when I click on the "Play" button so that it can start the first level. I then added the script to quit the game, this doesnt quit unity though. Debug was added to both of these buttons for testing purposes.