Eleventh Commit
I added a tilemap which consisted of 31 tiles of 16x16 and splitted them so they were all single. I made sure to turn off compression and make it point to point filter to stop it from being blury. A tilemap was added to the game then I placed my tilemap into the palette where I added two different types of tilemaps one called "Tilemap Base" this is what the player will walk on it will have a collider on it, its layer will be set to "Ground". The other is "Tilemap Detail" this will be tiles such as flowers and will have no colliders upon them, its layer will be set to "Detail".I will possibly add to types of "Tilemap Detail" so that some of the details will be infront of the player and some will be behind the player. A "Tilemap Collider 2D" which used by composit was turned on, was added to "Tilemap Base" aaswell as "Composite Collider 2D" which also added a "Rigidbody 2D" which I set the body type to static. Otherwise tilemap will have gravity applied onto it and will fall out of the camera view. I will design the level next.