Twelfth Commit
I will now begin the level design I first duplicated the "Test Scene" by clicking "ctrl + D" and renamed one to "Level One". Im keeping the "Test Scene" to test new scripts, sprites and other assets. Next I deleted the platforms and move the boarders and void to the side so that I could have an empty scene to build my level with the tilemap. I spread the void along the bottom of the tilemap so if the player falls off they will then respawn with the "FallenOff" script. I then decreased the size of the "Box Collider 2D" on the player so that he could fall between the gaps in the bridge made by the tile map. I also added a "Physic Material 2D" called "Slippery Soap" which is set to 0 friction meaning the player will now slip off the side on the tilemap rather than the player gripping on the side of it and not falling off into the void. Then I moved both boarder to the left and right and also placed the empty named "Next Level" which has the "NextLevel" script.