Thirteenth Commit
I have decided to change how the game work rather than the player falling off and respawning theyll lose 1 heart/health and once the player has 0 hearts/health the player will be asked to restart the whole game which will direct the player to a scene where there will be a quit button and a retry button which will start the player all the way back at level one to make the game harder. Maybe the player will gain a new full set of hearts/health when they start a new level or pass a level. After changing the code for the script "FallenOff" I noticed that the player would take damage but not respawn back to the start of the level this was an easy fix using the players postition adn moving it to the empty named "Respawn Point" postion. In the "PlayerHealth" script I also removed the two public voids "Die" and "Resapwn" as the player will not repsawn without going throught the death menu and also becasue once the payer dies it will also have to go through the death menu.