Fourteenth Commit
I wanted to add movement to the camera when the player took damage but I couldnt animate it as the camera is set to follow the player meaning I couldnt do it unless I had a fixed camera position which I didnt want. So I did some reseacrh and found a shader pack which put a ripple in the camera and looks really good. All I had to do was copy the shader code and then place it on a material and then make a new script which I called "RippleProcessor" and added the material to it and then made a link to the "PlayerHealth" script so that when the player takes damage the ripple effect will occur.