Fifteenth Commit
I wanted text on screen for when the player walking into a collider which would be attached to nothing or an npc to tell a story etc. I made two scripts for this a "SomeoneTalking" script which is attached to a collider and I also made a "Talking" script this will probably be used for the start of the game to introuduce the player to the world as the script runs from a void start meaning as soon as the scene that its attached to starts the script will play, but sometimes we dont want this so I made the other script for those times. I made the sentences a public string array as I dont want to make a different script for every different collide or start, this way I can apply the same script and write whatever in the inspector. I also found out how to add a delay to the sentence to make it seem like it typing infront of the player with one letter coming on at a different time to each other one per a few milliseconds. I also added a button so that you can click it to go to the next sentence.