Seventeenth Commit
Comments were added to most of my scripts, whch were descriptions of what each funtion does whne its called ect. Some scripts were even deleted to clean up the assets in the game. For example, the "TalkingTrig" script and the "SomeoneTalking" script. In the "Talking" script I added the "PlayerController" to script and to disable it when text is displayed so that the player freezes when text is shown. I will also need to design and finish off the next "Level Two" and "Level Three".